securechat-app-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==1.0.3
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
liboqs-python>=0.8.0